synapta-table-segmentation/requirements.txt
# PDF Processing
pdfplumber>=0.10.0
pymupdf>=1.23.0
camelot-py[cv]>=0.11.0
tabula-py>=2.5.0

# OCR (fallback for raster tables)
pytesseract>=0.3.10
Pillow>=10.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# LLM/Embeddings (for descriptions and semantic matching)
openai>=1.0.0  # Optional: for LLM-assisted descriptions
sentence-transformers>=2.2.0  # For semantic similarity

# Utilities
python-dateutil>=2.8.0
regex>=2023.0.0






